requirements.txt
google-generativeai
openai
fastmcp==2.11.3
pyodbc
pywin32
python-telegram-bot

